Removing the classes/config.class.php file from repo
I'm removing the classes/config.class.php file from the repo as it should
probably not be "tracked" so to speak. Any necessary changes should be made
in the classes/config.class.sample.php, and installation specific changes
can be made in the other one.

By removing it from the repo, we can set up our local copies without risking
the passwords being pushed out on the internet. After all, we won't be changing
those values in the configuration anytime soon.